+++ b2sdk/transfer/outbound/upload_source.py
@@ -10,15 +10,29 @@
 
 import hashlib
 import io
+import logging
 import os
 
 from abc import abstractmethod
-from typing import Callable, Optional, Union
+from enum import auto, Enum, unique
+from typing import Callable, List, Optional, Union
 
 from b2sdk.exception import InvalidUploadSource
+from b2sdk.file_version import BaseFileVersion
+from b2sdk.http_constants import DEFAULT_MIN_PART_SIZE
 from b2sdk.stream.range import RangeOfInputStream, wrap_with_range
+from b2sdk.transfer.outbound.copy_source import CopySource
 from b2sdk.transfer.outbound.outbound_source import OutboundTransferSource
-from b2sdk.utils import hex_sha1_of_stream, hex_sha1_of_unlimited_stream, Sha1HexDigest
+from b2sdk.utils import hex_sha1_of_unlimited_stream, Sha1HexDigest, IncrementalHexDigester, hex_sha1_of_stream
+
+logger = logging.getLogger(__name__)
+
+
+@unique
+class UploadMode(Enum):
+    """ Mode of file uploads """
+    FULL = auto()  #: always upload the whole file
+    INCREMENTAL = auto()  #: use incremental uploads when possible
 
 
 class AbstractUploadSource(OutboundTransferSource):
@@ -68,7 +82,7 @@ class UploadSourceBytes(AbstractUploadSource):
         self.data_bytes = data_bytes
         self.content_sha1 = content_sha1
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         return '<{classname} data={data} id={id}>'.format(
             classname=self.__class__.__name__,
             data=str(self.data_bytes[:20]) +
@@ -76,10 +90,10 @@ class UploadSourceBytes(AbstractUploadSource):
             id=id(self),
         )
 
-    def get_content_length(self):
+    def get_content_length(self) -> int:
         return len(self.data_bytes)
 
-    def get_content_sha1(self):
+    def get_content_sha1(self) -> Optional[Sha1HexDigest]:
         if self.content_sha1 is None:
             self.content_sha1 = hashlib.sha1(self.data_bytes).hexdigest()
         return self.content_sha1
@@ -87,11 +101,11 @@ class UploadSourceBytes(AbstractUploadSource):
     def open(self):
         return io.BytesIO(self.data_bytes)
 
-    def is_sha1_known(self):
+    def is_sha1_known(self) -> bool:
         return self.content_sha1 is not None
 
 
-class UploadSourceLocalFile(AbstractUploadSource):
+class UploadSourceLocalFileBase(AbstractUploadSource):
     def __init__(
         self,
         local_path: Union[os.PathLike, str],
@@ -105,16 +119,15 @@ class UploadSourceLocalFile(AbstractUploadSource):
         """
         self.local_path = local_path
         self.content_length = 0
+        self.content_sha1 = content_sha1
         self.check_path_and_get_size()
 
-        self.content_sha1 = content_sha1
-
-    def check_path_and_get_size(self):
+    def check_path_and_get_size(self) -> None:
         if not os.path.isfile(self.local_path):
             raise InvalidUploadSource(self.local_path)
         self.content_length = os.path.getsize(self.local_path)
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         return (
             '<{classname} local_path={local_path} content_length={content_length} '
             'content_sha1={content_sha1} id={id}>'
@@ -126,10 +139,10 @@ class UploadSourceLocalFile(AbstractUploadSource):
             id=id(self),
         )
 
-    def get_content_length(self):
+    def get_content_length(self) -> int:
         return self.content_length
 
-    def get_content_sha1(self):
+    def get_content_sha1(self) -> Optional[Sha1HexDigest]:
         if self.content_sha1 is None:
             self.content_sha1 = self._hex_sha1_of_file()
         return self.content_sha1
@@ -137,15 +150,15 @@ class UploadSourceLocalFile(AbstractUploadSource):
     def open(self):
         return io.open(self.local_path, 'rb')
 
-    def _hex_sha1_of_file(self):
+    def _hex_sha1_of_file(self) -> Sha1HexDigest:
         with self.open() as f:
             return hex_sha1_of_stream(f, self.content_length)
 
-    def is_sha1_known(self):
+    def is_sha1_known(self) -> bool:
         return self.content_sha1 is not None
 
 
-class UploadSourceLocalFileRange(UploadSourceLocalFile):
+class UploadSourceLocalFileRange(UploadSourceLocalFileBase):
     def __init__(
         self,
         local_path: Union[os.PathLike, str],
@@ -162,7 +175,7 @@ class UploadSourceLocalFileRange(UploadSourceLocalFile):
         :param length: Amount of data to be uploaded. If ``None``, length of
                       the remainder of the file is taken.
         """
-        super(UploadSourceLocalFileRange, self).__init__(local_path, content_sha1)
+        super().__init__(local_path, content_sha1)
         self.file_size = self.content_length
         self.offset = offset
         if length is None:
@@ -172,7 +185,7 @@ class UploadSourceLocalFileRange(UploadSourceLocalFile):
                 raise ValueError('Range length overflow file size')
             self.content_length = length
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         return (
             '<{classname} local_path={local_path} offset={offset} '
             'content_length={content_length} content_sha1={content_sha1} id={id}>'
@@ -188,6 +201,84 @@ class UploadSourceLocalFileRange(UploadSourceLocalFile):
     def open(self):
         fp = super(UploadSourceLocalFileRange, self).open()
         return wrap_with_range(fp, self.file_size, self.offset, self.content_length)
+
+
+class UploadSourceLocalFile(UploadSourceLocalFileBase):
+    def get_incremental_sources(
+        self,
+        file_version: BaseFileVersion,
+        min_part_size: Optional[int] = None,
+    ) -> List[OutboundTransferSource]:
+        """
+        Split the upload into a copy and upload source constructing an incremental upload
+
+        This will return a list of upload sources.  If the upload cannot split, the method will return [self].
+        """
+
+        if not file_version:
+            logger.debug(
+                "Fallback to full upload for %s -- no matching file on server", self.local_path
+            )
+            return [self]
+
+        min_part_size = min_part_size or DEFAULT_MIN_PART_SIZE
+        if file_version.size < min_part_size:
+            # existing file size below minimal large file part size
+            logger.debug(
+                "Fallback to full upload for %s -- remote file is smaller than %i bytes",
+                self.local_path, min_part_size
+            )
+            return [self]
+
+        if self.get_content_length() < file_version.size:
+            logger.debug(
+                "Fallback to full upload for %s -- local file is smaller than remote",
+                self.local_path
+            )
+            return [self]
+
+        content_sha1 = file_version.get_content_sha1()
+
+        if not content_sha1:
+            logger.debug(
+                "Fallback to full upload for %s -- remote file content SHA1 unknown",
+                self.local_path
+            )
+            return [self]
+
+        # We're calculating hexdigest of the first N bytes of the file. However, if the sha1 differs,
+        # we'll be needing the whole hash of the file anyway. So we can use this partial information.
+        digester = IncrementalHexDigester()
+        with self.open() as fp:
+            hex_digest = digester.update_from_stream(fp, file_version.size)
+            if hex_digest != content_sha1:
+                logger.debug(
+                    "Fallback to full upload for %s -- content in common range differs",
+                    self.local_path,
+                )
+                # Calculate SHA1 of the remainder of the file and set it.
+                self.content_sha1 = digester.update_from_stream(fp)
+                return [self]
+
+        logger.debug("Incremental upload of %s is possible.", self.local_path)
+
+        if file_version.server_side_encryption and file_version.server_side_encryption.is_unknown():
+            source_encryption = None
+        else:
+            source_encryption = file_version.server_side_encryption
+
+        sources = [
+            CopySource(
+                file_version.id_,
+                offset=0,
+                length=file_version.size,
+                encryption=source_encryption,
+                source_file_info=file_version.file_info,
+                source_content_type=file_version.content_type,
+            ),
+            UploadSourceLocalFileRange(self.local_path, offset=file_version.size),
+        ]
+        return sources
 
 
 class UploadSourceStream(AbstractUploadSource):
@@ -210,7 +301,7 @@ class UploadSourceStream(AbstractUploadSource):
         self._content_length = stream_length
         self._content_sha1 = stream_sha1
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         return (
             '<{classname} stream_opener={stream_opener} content_length={content_length} '
             'content_sha1={content_sha1} id={id}>'
@@ -222,12 +313,12 @@ class UploadSourceStream(AbstractUploadSource):
             id=id(self),
         )
 
-    def get_content_length(self):
+    def get_content_length(self) -> int:
         if self._content_length is None:
             self._set_content_length_and_sha1()
         return self._content_length
 
-    def get_content_sha1(self):
+    def get_content_sha1(self) -> Optional[Sha1HexDigest]:
         if self._content_sha1 is None:
             self._set_content_length_and_sha1()
         return self._content_sha1
@@ -235,12 +326,12 @@ class UploadSourceStream(AbstractUploadSource):
     def open(self):
         return self.stream_opener()
 
-    def _set_content_length_and_sha1(self):
+    def _set_content_length_and_sha1(self) -> None:
         sha1, content_length = hex_sha1_of_unlimited_stream(self.open())
         self._content_length = content_length
         self._content_sha1 = sha1
 
-    def is_sha1_known(self):
+    def is_sha1_known(self) -> bool:
         return self._content_sha1 is not None
 
 
@@ -269,7 +360,7 @@ class UploadSourceStreamRange(UploadSourceStream):
         )
         self._offset = offset
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         return (
             '<{classname} stream_opener={stream_opener} offset={offset} '
             'content_length={content_length} content_sha1={content_sha1} id={id}>'